+++ src/objects/sphere_utils.c
@@ -22,11 +22,14 @@ t_point3	ft_collide_sphere(void *fig, t_point3 origin, t_point3 direct)
 	sph = (t_sphere *)fig;
 	if (ft_3_line_point_dist(origin, direct, sph->origin) > sph->radius)
 		return (ft_3_nullpointnew());
-	ft_solve_sqr((pow(direct.x, 2) + pow(direct.y, 2) + pow(direct.z, 2)), 2.0 *
-		(direct.x * (origin.x - sph->origin.x) + direct.y * (origin.y -
-			sph->origin.y) + direct.z * (origin.z - sph->origin.z)),
-		(pow(origin.x - sph->origin.x, 2) + pow(origin.y - sph->origin.y, 2) +
-		pow(origin.z - sph->origin.z, 2) - pow(sph->radius, 2)), &res);
+	ft_solve_sqr(
+		(float)(pow(direct.x, 2) + pow(direct.y, 2) + pow(direct.z, 2)),
+		2.0f * (direct.x * (origin.x - sph->origin.x) +
+			direct.y * (origin.y - sph->origin.y) +
+			direct.z * (origin.z - sph->origin.z)),
+		(float)(pow(origin.x - sph->origin.x, 2) +
+			pow(origin.y - sph->origin.y, 2) +
+			pow(origin.z - sph->origin.z, 2) - pow(sph->radius, 2)), &res);
 	if (!res[0])
 		return (ft_3_nullpointnew());
 	coll_points[0] = (res[1] > 0) ? ft_3_vector_add(origin,
